Abstracted out app.js from start.js; double checked .uses Still works (NB kernel upgrade here)
