+++ server/start.js
@@ -1,14 +1,6 @@
-const express = require('express')
-const app = express()
-
-var someJSON = {data:7};
-var someJstring = JSON.stringify (someJSON);
-
-
-app.get('/', (req, res) => res.send('Hello World!'));
-app.get('/api', (req, res) => {
-    res.send (someJstring)
-  });
+const express = require('express');
+var path = require('path');
+var app = require('./app');
 
 const PORT = 8000;
